requirements.txt
# Django and core packages
Django>=5.2
djangorestframework>=3.14.0
channels>=4.0.0
channels-redis>=4.1.0
django-cors-headers>=4.3.0

# PostgreSQL adapter
psycopg2-binary>=2.9.9

# Solar controller components
pymodbus>=3.5.0

# Other dependencies may be listed here
